Continuar editar tarefa

+++ routes/web.php
@@ -23,9 +23,12 @@ Route::get('/', function () {
 // Route Task
 Route::get('tasks', [TaskController::class, 'index'])->name('tasks')->middleware('auth');
 Route::get('/cadTask', [TaskController::class, 'create'])->name('cadtask')->middleware('auth');
+Route::get('/editTask/{task}', [TaskController::class, 'edit'])->name('edittask')->middleware('auth');
+Route::post('/cadTask', [TaskController::class, 'store'])->name('instask')->middleware('auth');
+Route::put('/editTask/{task}', [TaskController::class, 'update'])->name('updatetask')->middleware('auth');
+Route::delete('/Tasks/destroy/{task}', [TaskController::class, 'destroy'])->name('deltask')->middleware('auth');;
 
 //Route Task Autocomplete
-//Route::get('/categoryAutocomplete', [TaskController::class, 'autocomplete'])->name('categoryautocomplete')->middleware('auth');
 Route::post('/getautocomplete', [TaskController::class, 'getautocomplete'])->name('getautocomplete')->middleware('auth');
 
 // Route User
@@ -33,7 +36,7 @@ Route::get('/users', [UserController::class, 'index'])->name('users')->middlewar
 Route::get('/cadUser', function () { if(Auth::user()->mod_id === 1) { return view('cadUser'); } })->name('caduser')->middleware('auth');
 Route::get('/editUser/{user}', [UserController::class, 'edit'])->name('edituser')->middleware('auth');
 Route::put('/editUser/{user}', [UserController::class, 'update'])->name('updateuser')->middleware('auth');
-Route::delete('/users/destroy/{user}', [UserController::class, 'destroy'])->name('delUsers');
+Route::delete('/users/destroy/{user}', [UserController::class, 'destroy'])->name('delUsers')->middleware('auth');;
 
 // Route Categories
 Route::get('/categories', [CategoryController::class, 'index'])->name('categories')->middleware('auth');
@@ -41,6 +44,6 @@ Route::get('/cadCategory', [CategoryController::class, 'create'])->name('cadcate
 Route::get('/editCategory/{category}', [CategoryController::class, 'edit'])->name('editcategory')->middleware('auth');
 Route::post('/cadCategory', [CategoryController::class, 'store'])->name('inscategory')->middleware('auth');
 Route::put('/editCategory/{category}', [CategoryController::class, 'update'])->name('updatecategory')->middleware('auth');
-Route::delete('/category/destroy/{category}', [CategoryController::class, 'destroy'])->name('delcategory');
+Route::delete('/category/destroy/{category}', [CategoryController::class, 'destroy'])->name('delcategory')->middleware('auth');;
 
 require __DIR__.'/auth.php';
